Extract ListInput
